fix: refactor to have inspect_helper.go
