+++ cmd/notation/internal/display/metadata/tree/blob_inspect.go
@@ -35,8 +35,8 @@ func NewBlobInspectHandler(printer *output.Printer) *BlobInspectHandler {
 }
 
 // OnEnvelopeParsed sets the parsed envelope for the handler.
-func (h *BlobInspectHandler) OnEnvelopeParsed(signaturePath, signatureEnvelopeType string, envelope signature.Envelope) error {
-	sigNode, err := newSignatureNode(signaturePath, signatureEnvelopeType, envelope)
+func (h *BlobInspectHandler) OnEnvelopeParsed(signaturePath, signatureMediaType string, envelope signature.Envelope) error {
+	sigNode, err := newSignatureNode(signaturePath, signatureMediaType, envelope)
 	if err != nil {
 		return err
 	}
